move ability/move tests into /src/tests and rename to *.test.ts to match common naming patterns
